fix(qa): Skip network policy deletion for doc-only PRs

+++ Jenkinsfile
@@ -246,11 +246,13 @@ node {
     stage('Post') {
       kubeHelper.teardown(kubeLocks)
       testHelper.teardown()
-      // tear down network policies deployed by the tests
-      kubeHelper.kube(kubectlNamespace, {
+      if(!skipUnitTests) {
+        // tear down network policies deployed by the tests
+        kubeHelper.kube(kubectlNamespace, {
           sh(script: 'kubectl --namespace="' + kubectlNamespace + '" delete networkpolicies --all', returnStatus: true);
         });
-      pipelineHelper.teardown(currentBuild.result)
+        pipelineHelper.teardown(currentBuild.result)
+      }
     }
   }
 }
